api/requirements.txt
fastapi
uvicorn
paho-mqtt
requests
websockets
asyncio